refactoring slashing risk actions
