+++ validator/client/src/main/java/tech/pegasys/teku/validator/client/restapi/apis/PostKeys.java
@@ -31,23 +31,23 @@ import tech.pegasys.teku.infrastructure.restapi.endpoints.EndpointMetadata;
 import tech.pegasys.teku.infrastructure.restapi.endpoints.RestApiEndpoint;
 import tech.pegasys.teku.infrastructure.restapi.endpoints.RestApiRequest;
 import tech.pegasys.teku.validator.client.KeyManager;
-import tech.pegasys.teku.validator.client.doppelganger.DoppelgangerDetectionAction;
 import tech.pegasys.teku.validator.client.doppelganger.DoppelgangerDetector;
 import tech.pegasys.teku.validator.client.restapi.ValidatorTypes;
 import tech.pegasys.teku.validator.client.restapi.apis.schema.PostKeysRequest;
+import tech.pegasys.teku.validator.client.slashingriskactions.SlashingRiskAction;
 
 public class PostKeys extends RestApiEndpoint {
   private final KeyManager keyManager;
   public static final String ROUTE = "/eth/v1/keystores";
   private final Path slashingProtectionPath;
   private final Optional<DoppelgangerDetector> maybeDoppelgangerDetector;
-  private final DoppelgangerDetectionAction doppelgangerDetectionAction;
+  private final SlashingRiskAction doppelgangerDetectionAction;
 
   public PostKeys(
       final KeyManager keyManager,
       final Path slashingProtectionPath,
       final Optional<DoppelgangerDetector> maybeDoppelgangerDetector,
-      final DoppelgangerDetectionAction doppelgangerDetectionAction) {
+      final SlashingRiskAction doppelgangerDetectionAction) {
     super(
         EndpointMetadata.post(ROUTE)
             .operationId("ImportKeystores")
