+++ validator/client/src/main/java/tech/pegasys/teku/validator/client/slashingriskactions/DoppelgangerDetectionShutDown.java
@@ -11,28 +11,21 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.validator.client.doppelganger;
+package tech.pegasys.teku.validator.client.slashingriskactions;
+
+import static tech.pegasys.teku.infrastructure.logging.StatusLogger.STATUS_LOG;
 
 import java.util.List;
 import java.util.stream.Collectors;
 import tech.pegasys.teku.bls.BLSPublicKey;
-import tech.pegasys.teku.infrastructure.logging.StatusLogger;
 
-public class DoppelgangerDetectionAlert implements DoppelgangerDetectionAction {
-
-  private final StatusLogger statusLog;
-
-  public DoppelgangerDetectionAlert() {
-    this(StatusLogger.STATUS_LOG);
-  }
-
-  public DoppelgangerDetectionAlert(final StatusLogger statusLog) {
-    this.statusLog = statusLog;
-  }
-
+public class DoppelgangerDetectionShutDown implements SlashingRiskAction {
   @Override
   public void perform(final List<BLSPublicKey> doppelgangers) {
-    statusLog.doppelgangerDetectionAlert(
-        doppelgangers.stream().map(BLSPublicKey::toAbbreviatedString).collect(Collectors.toSet()));
+    STATUS_LOG.exitOnDoppelgangerDetected(
+        doppelgangers.stream()
+            .map(BLSPublicKey::toAbbreviatedString)
+            .collect(Collectors.joining(", ")));
+    shutdown();
   }
 }
