+++ validator/client/src/main/java/tech/pegasys/teku/validator/client/slashingriskactions/SlashingRiskAction.java
@@ -11,12 +11,16 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.validator.client.doppelganger;
+package tech.pegasys.teku.validator.client.slashingriskactions;
 
 import java.util.List;
 import tech.pegasys.teku.bls.BLSPublicKey;
 
-public interface DoppelgangerDetectionAction {
+public interface SlashingRiskAction {
 
   void perform(final List<BLSPublicKey> doppelgangers);
+
+  default void shutdown() {
+    System.exit(2);
+  }
 }
